% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Categorize dataframe columns}
\usage{
categorize(df, max_cat = 10)
}
\arguments{
\item{df}{a data.frame}

\item{max_cat}{int, the maximum number of unique values that
define a categorical column}
}
\value{
list with sub vectors named 'numeric' and 'categorical',
containing column names of each type
}
\description{
Denote each dataframe category as numeric, categorical, or text
and return a list of lists labeled 'numeric',
'categorical' and 'text' containing column names
that fall into each category
}
\examples{
categorize(data.frame(a = c(1.2,2.3,3.4), b = c('a','b','c')))
}
