% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection.R
\name{feature_selection}
\alias{feature_selection}
\title{Apply feature selection for the input data}
\usage{
feature_selection(X, y, mode, n_features)
}
\arguments{
\item{X}{data.frame of predictors}

\item{y}{vector of response and should be a factor in case of classification}

\item{mode}{string regression or classification}

\item{n_features}{int}
}
\value{
list
}
\description{
Take input as predictons and response and return the
list of features that are important
}
\examples{
feature_selection(data.frame(X1 = c(2, 4, 3), X2 = c(8, 7, 4)), c(4, 3, 5), "regression", 2)
}
