% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing.R
\name{fill_missing}
\alias{fill_missing}
\title{Handles missing values in a dataframe}
\usage{
fill_missing(x_train, x_test, column_list, num_imp, cat_imp)
}
\arguments{
\item{x_train}{training set dataframe to be transformed}

\item{x_test}{test set dataframe to be transformed}

\item{column_list}{named list of columns with two sub vectors, must be named
'numeric' and 'categorical'}

\item{num_imp}{method for numerical imputation (default = "mean")}

\item{cat_imp}{method for categorical imputation}
}
\value{
list, with missing values replaced by the specified method
}
\description{
Replace missing values in dataframe columns by the specified methods.
Separate methods can be applied for categorical column imputation and
numerical column imputation.
}
\examples{
fill_missing(x_tr, x_test, list("numeric" = c('x'), "categorical" = c('y')), 'mean', 'mode')

}
