% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Categorize dataframe columns}
\usage{
categorize(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
list
}
\description{
Denote each dataframe category as numeric, categorical, or text
and return a list of lists labeled 'numeric', 
'categorical' and 'text' containing column names
that fall into each category
}
\examples{
categorize(data.frame(a = c(1.2,2.3), b = c('a','b','c')))
}
