% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_transformer.R
\name{column_transformer}
\alias{column_transformer}
\title{Transforms categorical and numerical features}
\usage{
column_transformer(df, num_list, cat_list, num_trans, cat_trans)
}
\arguments{
\item{df}{dataframe to be transformed}

\item{num_list}{list of columns to have numerical transformation   to be applied}

\item{cat_list}{list of columns to have categorical transformations to be applied}

\item{num_trans}{method for numerical transformation (default = "standard scaling")}

\item{cat_trans}{method for categorical transformation (default = "ohe")}
}
\value{
df, with tranformed column values
}
\description{
Transforms categorical and numerical features
}
\examples{
column_transformer(df, list('weight'), list('education'))

}
