% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_transformer.R
\name{column_transformer}
\alias{column_transformer}
\title{Transforms columns of a dataframe}
\usage{
column_transformer(
  x_train,
  x_test,
  column_list,
  num_trans = "standard_scaling",
  cat_trans = "onehot_encoding"
)
}
\arguments{
\item{x_train}{training set dataframe/tibble}

\item{x_test}{test set dataframe/tibble}

\item{column_list}{named list of categorical and numeric columns.}

\item{num_trans}{method(character) for numerical transformation - Can take values "standard_scaling" or "minmax_scaling" (default = "standard_scaling")}

\item{cat_trans}{method(character) for categorical transformation - Cant take values "onehot_encoding" or "label_encoding" (default = "onehot_encoding")}
}
\value{
A list with named items x_train and x_list that have been transformed according to the arguments specified
}
\description{
Transforms  columns in dataframe  by the specified methods.
Separate methods can be applied for categorical column transformation and
numerical column transformation.
}
\examples{
x_train <- data.frame('x' = c(2.5, 3.3, 5), 'y' = c(1, 6, 1))
x_test <- data.frame('x' = c(2), 'y' = c(1))
column_transformer(x_train, x_test, list("numeric" = c('x'), "categorical" = c('y')))
}
