% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_transformer.R
\name{column_transformer}
\alias{column_transformer}
\title{Transforms columns of a dataframe}
\usage{
column_transformer(
  x_train,
  x_test,
  col_names,
  num_trans = "standard_scaling",
  cat_trans = "onehot_encoding"
)
}
\arguments{
\item{x_train}{trainingset dataframe/tibble}

\item{x_test}{test set dataframe/tibble}

\item{col_names}{list of names of categorical columns}

\item{num_trans}{method(character) for numerical transformation (default = "standard_scaling")}

\item{cat_trans}{method(character) for categorical transformation (default = "ohe")}
}
\value{
list(x_train,x_test) transformed
}
\description{
Transforms  columns in dataframe  by the specified methods.
Separate methods can be applied for categorical column transformation and
numerical column transformation.
}
