% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing.R
\name{fill_missing}
\alias{fill_missing}
\title{Handles missing values in a dataframe}
\usage{
fill_missing(df, num_list, cat_list, num_imp, cat_imp)
}
\arguments{
\item{df}{dataframe to be transformed}

\item{num_list}{list of columns to have numerical imputation applied}

\item{cat_list}{list of columns to have categorical imputation applied}

\item{num_imp}{method for numerical imputation}

\item{cat_imp}{method for categorical imputation}
}
\value{
df, with missing values replaced by the specified method
}
\description{
Replace missing values in dataframe columns by the specified methods. 
Separate methods can be applied for categorical column imputation and 
numerical column imputation.
}
\examples{
fill_missing(df, list('weight'), list('education'), 'mean', 'mode')

}
